backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.5.0
python-dotenv==1.0.0
pymongo==4.6.0
httpx==0.25.2
google-generativeai==0.3.2
